Fixes #114 New SOLR 5.2 integration and new spatial queries

+++ web-tg/src/main/java/com/nextep/proto/action/impl/tools/IndexActivitiesAction.java
@@ -8,8 +8,13 @@ import org.apache.commons.logging.LogFactory;
 
 import com.nextep.activities.model.Activity;
 import com.nextep.activities.model.ActivityType;
+import com.nextep.geo.model.GeographicItem;
+import com.nextep.media.model.Media;
 import com.nextep.proto.action.base.AbstractAction;
+import com.nextep.proto.apis.adapters.ApisActivityExtraKeyAdapter;
 import com.nextep.proto.apis.adapters.ApisActivityTargetKeyAdapter;
+import com.nextep.proto.apis.adapters.ApisEventLocationAdapter;
+import com.nextep.proto.helpers.GeoHelper;
 import com.nextep.proto.spring.ContextHolder;
 import com.nextep.smaug.service.SearchPersistenceService;
 import com.videopolis.apis.exception.ApisException;
@@ -18,6 +23,8 @@ import com.videopolis.apis.factory.SearchRestriction;
 import com.videopolis.apis.model.ApisCriterion;
 import com.videopolis.apis.model.ApisRequest;
 import com.videopolis.apis.service.ApiCompositeResponse;
+import com.videopolis.calm.exception.CalException;
+import com.videopolis.calm.model.CalmObject;
 import com.videopolis.cals.factory.ContextFactory;
 import com.videopolis.cals.model.PaginationInfo;
 import com.videopolis.concurrent.exception.TaskExecutionException;
@@ -33,11 +40,15 @@ public class IndexActivitiesAction extends AbstractAction {
 			.getLog(IndexActivitiesAction.class);
 	private static final long serialVersionUID = 8242078679018166793L;
 	private static final String APIS_ALIAS_ACTIVITIES = "activities";
+	private static final String APIS_ALIAS_ACTIVITY_OBJECT = "object";
+	private static final String APIS_ALIAS_EVENT_LOCATION = "eventloc";
+	private static final String APIS_ALIAS_ACTIVITY_EXTRA = "extra";
 	private static final int PAGE_SIZE = 500;
 	private SearchPersistenceService searchService;
 	private List<String> messages = new ArrayList<String>();
 	private TaskRunnerService taskRunnerService;
 	private String type = null;
+	private boolean clearActivities = false;
 
 	@Override
 	protected String doExecute() throws Exception {
@@ -47,12 +58,7 @@ public class IndexActivitiesAction extends AbstractAction {
 					throws TaskExecutionException, InterruptedException {
 				try {
 					indexActivities();
-				} catch (ApisException e) {
-					LOGGER.error(
-							"Problem while indexing cities : " + e.getMessage(),
-							e);
-					throw new TaskExecutionException(e);
-				} catch (RuntimeException e) {
+				} catch (ApisException | CalException | RuntimeException e) {
 					LOGGER.error(
 							"Problem while indexing cities : " + e.getMessage(),
 							e);
@@ -65,10 +71,14 @@ public class IndexActivitiesAction extends AbstractAction {
 		return SUCCESS;
 	}
 
-	private void indexActivities() throws ApisException {
+	private void indexActivities() throws ApisException, CalException {
 		int offset = 0;
 		int item = 1;
 		boolean hasMore = true;
+		if (clearActivities) {
+			LOGGER.warn("REMOVING ALL ACTIVITIES");
+			searchService.removeAll(Activity.CAL_TYPE);
+		}
 		while (hasMore) {
 			final ApisRequest request = buildRequest(offset++);
 			ContextHolder.toggleReadonly();
@@ -89,19 +99,45 @@ public class IndexActivitiesAction extends AbstractAction {
 				LOGGER.info("Storing activity '" + activity.getKey().toString()
 						+ "' [" + item + " / " + itemCount + "]");
 				item++;
-				try {
-					if (type == null
-							|| (type != null && activity.getActivityType()
-									.name().equals(type))) {
-						if (activity.getActivityType() != ActivityType.LOCALIZATION) {
-							searchService.storeCalmObject(activity,
-									SearchScope.CHILDREN);
-						} else {
-							searchService.remove(activity);
-						}
+
+				// Extracting target object
+				final CalmObject targetObject = activity.getUnique(
+						CalmObject.class, APIS_ALIAS_ACTIVITY_OBJECT);
+				final CalmObject extraObject = activity.getUnique(
+						CalmObject.class, APIS_ALIAS_ACTIVITY_EXTRA);
+
+				if (isActivityIndexed(targetObject, extraObject)) {
+					GeographicItem geoObject = GeoHelper
+							.extractLocalization(targetObject);
+					// Trying to extract event location
+					if (geoObject == null) {
+						geoObject = activity.getUnique(GeographicItem.class,
+								APIS_ALIAS_EVENT_LOCATION);
 					}
-				} catch (SearchException e) {
-					LOGGER.error(" ==> FAILED : " + e.getMessage());
+					if (geoObject != null) {
+						activity.add(geoObject);
+					} else {
+						LOGGER.info("Unable to load location for '"
+								+ targetObject.getKey() + "'");
+					}
+					try {
+						if (type == null
+								|| (type != null && activity.getActivityType()
+										.name().equals(type))) {
+							if (activity.getActivityType() != ActivityType.LOCALIZATION) {
+								searchService.storeCalmObject(activity,
+										SearchScope.CHILDREN);
+							} else {
+								searchService.remove(activity);
+							}
+						}
+					} catch (SearchException e) {
+						LOGGER.error(" ==> FAILED : " + e.getMessage());
+					}
+				} else {
+					LOGGER.warn("Activity " + activity.getKey()
+							+ " target object key "
+							+ activity.getLoggedItemKey() + " not found");
 				}
 			}
 		}
@@ -116,8 +152,24 @@ public class IndexActivitiesAction extends AbstractAction {
 								.paginatedBy(PAGE_SIZE, offset)
 								.aliasedBy(APIS_ALIAS_ACTIVITIES)
 								.addCriterion(
+										(ApisCriterion) SearchRestriction
+												.adaptKey(
+														new ApisActivityTargetKeyAdapter())
+												.aliasedBy(
+														APIS_ALIAS_ACTIVITY_OBJECT)
+												.with(GeographicItem.class)
+												.addCriterion(
+														SearchRestriction
+																.adaptKey(
+																		new ApisEventLocationAdapter())
+																.aliasedBy(
+																		APIS_ALIAS_EVENT_LOCATION)))
+								.addCriterion(
 										SearchRestriction
-												.adaptKey(new ApisActivityTargetKeyAdapter())));
+												.adaptKey(
+														new ApisActivityExtraKeyAdapter())
+												.aliasedBy(
+														APIS_ALIAS_ACTIVITY_EXTRA)));
 	}
 
 	public void setSearchService(SearchPersistenceService searchService) {
@@ -138,5 +190,24 @@ public class IndexActivitiesAction extends AbstractAction {
 
 	public String getType() {
 		return type;
+	}
+
+	public void setClearActivities(boolean clearActivities) {
+		this.clearActivities = clearActivities;
+	}
+
+	public boolean isClearActivities() {
+		return clearActivities;
+	}
+
+	private boolean isActivityIndexed(CalmObject targetObj, CalmObject extraObj) {
+		if (targetObj == null) {
+			return false;
+		} else {
+			if (extraObj instanceof Media) {
+				return ((Media) extraObj).isOnline();
+			}
+		}
+		return true;
 	}
 }
